New files created, many methods moved around to address seperation of concerns. Several using statements changed and/or removed.
