+++ src/Chirp.Core/DTOs/AuthorDTO.cs
@@ -1,0 +1,1 @@
+namespace Chirp.Core.DTOs;
