+++ src/Chirp.Infrastructure/Services/AuthorService.cs
@@ -1,0 +1,41 @@
+using System.Data;
+using Chirp.Core.DTOs;
+using Chirp.Core.Entities;
+using Chirp.Infrastructure.Repositories;
+using Microsoft.Data.Sqlite;
+
+//always define namespace everywhere except for Program.cs
+namespace Chirp.Infrastructure.Services;
+
+public class AuthorService : IAuthorService
+{
+    private readonly IAuthorRepository _authorRepository;
+
+    public AuthorService(IAuthorRepository authorRepository)
+    {
+        _authorRepository = authorRepository;
+    }
+
+    public async Task<Author> CreateAuthor(string name, string email)
+    {
+        // Fetch the latest ID as a string
+        var latestIdString = await _authorRepository.GetLatestIdAuthor();
+
+        // Parse the ID to an integer 
+        var newId = (int.Parse(latestIdString) + 1).ToString();
+
+        var author = new Author()
+        {
+            Id = newId, // Assign the new incremented ID as a string
+            UserName = name,
+            Email = email,
+            Cheeps = new List<Cheep>()
+        };
+        return author;
+    }
+
+    public async Task AddAuthor(Author author)
+    {
+        await _authorRepository.AddAuthor(author);
+    }
+}
