+++ src/Chirp.Web/Areas/Identity/Pages/Account/AboutMe.cshtml.cs
@@ -6,7 +6,6 @@ using Chirp.Core.Entities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authentication;
 using Chirp.Infrastructure.Repositories;
-using Chirp.Core.Repositories;
 using Microsoft.EntityFrameworkCore;
 
 namespace Chirp.Web.Areas.Identity.Pages.Account
@@ -15,12 +14,13 @@ namespace Chirp.Web.Areas.Identity.Pages.Account
     public class AboutMeModel : PageModel
     {
         private readonly ChirpDBContext _dbContext;
-        private readonly ICheepRepository _cheepRepository;
+        private readonly IAuthorRepository _authorRepository;
 
-        public AboutMeModel(ChirpDBContext dbContext, ICheepRepository cheepRepository)
+
+        public AboutMeModel(ChirpDBContext dbContext, IAuthorRepository authorRepository)
         {
             _dbContext = dbContext;
-            _cheepRepository = cheepRepository;
+            _authorRepository = authorRepository;
         }
 
         public string? UserName { get; set; }
@@ -45,7 +45,7 @@ namespace Chirp.Web.Areas.Identity.Pages.Account
             UserCheeps = _dbContext.Cheeps.Where(c => c.AuthorId == userId).ToList();
 
             // Get following list 
-            Following = await _cheepRepository.GetFollowedUsers(userId);
+            Following = await _authorRepository.GetFollowedUsers(userId);
         }
 
         public async Task<IActionResult> OnPostForgetMeAsync()
@@ -55,14 +55,14 @@ namespace Chirp.Web.Areas.Identity.Pages.Account
 
             // Fetch the user from db
             var user = await _dbContext.Authors
-                .Include(a => a.Follows) 
+                .Include(a => a.Follows)
                 .FirstOrDefaultAsync(a => a.Id == userId);
             if (user == null) return NotFound();
 
             // Remove entries where the user is a follower
             if (user.Follows != null && user.Follows.Any())
             {
-                _dbContext.Entry(user).Collection(u => u.Follows).Load(); 
+                _dbContext.Entry(user).Collection(u => u.Follows).Load();
                 user.Follows.Clear(); // Remove all follows
             }
