+++ src/Chirp.Web/Areas/Identity/Pages/Account/ConfirmEmail.cshtml.cs
@@ -45,10 +45,10 @@ namespace Chirp.Web.Areas.Identity.Pages.Account
 
             code = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(code));
             var result = await _userManager.ConfirmEmailAsync(user, code);
-            StatusMessage = result.Succeeded ? "Thank you for confirming your email. You will be automatically redirected" 
+            StatusMessage = result.Succeeded ? "Thank you for confirming your email. You will be automatically redirected"
                                                 : "Error confirming your email.";
-                                                
-            
+
+
             return Page();
         }
     }
